+++ wallet-connect/package.json
