+++ wallet-connect/src/App.vue
@@ -1,13 +1,72 @@
 <template>
-  <HomePage />
+  <div class="header">
+    <button type="button" class="walletButton" @click="connectAlgoSigner">
+      Connect AlgoSigner
+    </button>
+    <p v-if="walletAddress">Address: {{ walletAddress }}</p>
+    <button type="button" class="walletButton" @click="handleLogOut">
+      Disconnect AlgoSigner
+    </button>
+  </div>
 </template>
 
 <script lang="ts">
 import { defineComponent } from "vue";
-import HomePage from "./pages/HomePage.vue";
+import { CHAIN_NAME } from "./config/algosigner.config";
+import { WalletType } from "./types/enum.types";
+import WalletStore from "./store/WalletStore";
+declare var AlgoSigner: any; // eslint-disable-line
 
 export default defineComponent({
   name: "App",
-  components: { HomePage },
+  data() {
+    return {
+      walletAddress: "",
+    };
+  },
+  setup() {
+    const walletStore = WalletStore();
+    return {
+      setWalletType: walletStore.setWalletType,
+    };
+  },
+  methods: {
+    async connectAlgoSigner() {
+      try {
+        const algoSignerResponse = await AlgoSigner.connect({
+          ledger: CHAIN_NAME,
+        });
+        this.setWalletType(WalletType.ALGOSIGNER);
+        console.log("Connected to AlgoSigner:", algoSignerResponse);
+        await this.getUserAccount();
+      } catch (e) {
+        console.error(e);
+      }
+    },
+    async getUserAccount() {
+      const userAccount = await AlgoSigner.accounts({
+        ledger: CHAIN_NAME,
+      });
+      if (userAccount && userAccount.length) {
+        this.walletAddress = userAccount[0].address;
+      }
+    },
+    handleLogOut() {
+      console.log("Wallet Disconnected");
+      this.walletAddress = "";
+      this.setWalletType(WalletType.NONE);
+    },
+  },
 });
 </script>
+<style scoped>
+.header {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+.walletButton {
+  width: 20em;
+  margin: 5em;
+}
+</style>
